registry/fix: carousel shifted to primitives (#11)
