--- apps/registry/src/packages/components/carousel/01/docs.md
@@ -1,6 +1,0 @@
----
-name: Logo Carousel
-tags: []
-description: This is an infinite logo Carousel developed using only tailwind. It has a direction attribute that controls the direction of the animation where the default is left.
-version_included: 0.2.0
----
