+++ apps/registry/src/packages/primitives/carousel/01/default-js/tailwind.config.js
