+++ apps/registry/src/packages/primitives/carousel/01/default-ts/Logo.tsx
