+++ apps/registry/src/packages/primitives/carousel/01/docs.md
@@ -1,0 +1,6 @@
+---
+name: Carousel
+tags: []
+description: This is an infinite Carousel developed using only tailwind. It has a direction attribute that controls the direction of the animation where the default is left. It also has a pauseOnHover attribute that pauses the animation when the user hovers over the carousel.
+version_included: 0.2.0
+---
