+++ apps/web/src/app/docs/(generatedDocs)/primitives/containers/page.mdx
@@ -1320,7 +1320,7 @@ export const Primitives_containers_07 = lazy(() => import("@registry/packages/pr
 ```default|js|container.jsx
 "use client";
 
-import useMousePosition from "@/packages/hooks/useMousePosition/01/default-js/useMousePosition";
+import useMousePosition from "@registry/packages/hooks/useMousePosition/01/default-js/useMousePosition";
 import { useEffect, useRef, useState } from "react";
 import { twMerge } from "tailwind-merge";
 
@@ -1386,7 +1386,7 @@ export default function Example() {
 ```default|ts|container.tsx
 "use client";
 
-import useMousePosition from "@/packages/hooks/useMousePosition/01/default-ts/useMousePosition";
+import useMousePosition from "@registry/packages/hooks/useMousePosition/01/default-ts/useMousePosition";
 import { useEffect, useRef, useState } from "react";
 import { twMerge } from "tailwind-merge";
